moved enemies.py
